Chat-bot/chatbot/requirements.txt
fastapi[standard]
pydantic-settings
uvicorn[standard]
requests
psycopg2-binary
sqlalchemy
langchain
langchain_google_genai